fundamental_master/requirements.txt
# 核心依賴
python-dotenv>=1.0.0

# 網頁爬蟲
selenium>=4.16.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
requests>=2.31.0

# AI 分析
openai>=1.6.0

# Telegram Bot
python-telegram-bot>=20.0

# Web API
Flask>=3.0.0
flask-cors>=4.0.0

# 報告生成
playwright>=1.40.0
jinja2>=3.1.0

# 資料處理
pandas>=2.0.0
numpy>=1.24.0

# FinMind (備用資料源)
FinMind>=1.5.0

# 工具
tqdm>=4.60.0
